requirements.txt
# Python dependencies for Voice to Notes

# Web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
jinja2>=3.1.3

# Database
sqlalchemy>=2.0.25

# AI
google-generativeai>=0.8.0

# Utils
python-dotenv>=1.0.0
